+++ test/perf/pt_order_book.cpp
@@ -28,7 +28,7 @@ int run_test(TypedOrderBook& order_book, TypedOrder** orders, clock_t end) {
     }
     ++count;
   } while (clock() < end);
-  return (pp_order - orders);
+  return int(pp_order - orders);
 }
 
 template <class TypedOrderBook>
@@ -82,7 +82,7 @@ bool build_and_run_test(uint32_t dur_sec, uint32_t num_to_try) {
     std::cout << "Inserted " << count << " orders in " << dur_sec << " seconds"
               << ", or " << count / dur_sec << " insertions per sec"
               << std::endl;
-    uint32_t remain = order_book.bids().size() + order_book.asks().size();
+    uint32_t remain = uint32_t(order_book.bids().size() + order_book.asks().size());
     std::cout << "Run matched " << count - remain << " orders" << std::endl;
     return true;
   } else {
